requirements.txt
openai
langchain
langchain-community
chromadb
streamlit
tiktoken
python-dotenv
beautifulsoup4
requests